+++ rmk-macro/src/bind_interrupt.rs
@@ -3,7 +3,7 @@
 
 use proc_macro2::TokenStream as TokenStream2;
 use quote::{format_ident, quote};
-use rmk_config::toml_config::BleConfig;
+use crate::config::BleConfig;
 use syn::ItemMod;
 
 use crate::keyboard_config::KeyboardConfig;
@@ -61,6 +61,8 @@ pub(crate) fn bind_interrupt_default(keyboard_config: &KeyboardConfig) -> TokenS
                     battery_adc_pin: Some(_adc_pin),
                     charge_state: _,
                     charge_led: _,
+                    adc_divider_measured: _,
+                    adc_divider_total: _,
                 }) = keyboard_config.communication.get_ble_config()
                 {
                     Some(quote! {