+++ rmk-macro/src/lib.rs
@@ -1,7 +1,9 @@
+mod behavior;
 mod bind_interrupt;
 mod ble;
 mod chip_init;
 mod comm;
+mod config;
 mod default_config;
 mod entry;
 mod feature;