+++ rmk/src/config/mod.rs
@@ -8,6 +8,7 @@ pub use esp_config::BleBatteryConfig;
 #[cfg(feature = "_nrf_ble")]
 pub use nrf_config::BleBatteryConfig;
 
+use embassy_time::Duration;
 use embedded_hal::digital::OutputPin;
 
 /// Internal configurations for RMK keyboard.
@@ -17,6 +18,7 @@ pub struct RmkConfig<'a, O: OutputPin> {
     pub vial_config: VialConfig<'a>,
     pub light_config: LightConfig<O>,
     pub storage_config: StorageConfig,
+    pub behavior_config: BehaviorConfig,
     #[cfg(feature = "_nrf_ble")]
     pub ble_battery_config: BleBatteryConfig<'a>,
     #[cfg(feature = "_esp_ble")]
@@ -31,8 +33,29 @@ impl<'a, O: OutputPin> Default for RmkConfig<'a, O> {
             vial_config: VialConfig::default(),
             light_config: LightConfig::default(),
             storage_config: StorageConfig::default(),
+            behavior_config: BehaviorConfig::default(),
             #[cfg(any(feature = "_nrf_ble", feature = "_esp_ble"))]
             ble_battery_config: BleBatteryConfig::default(),
+        }
+    }
+}
+
+/// Config for configurable action behavior
+#[derive(Default)]
+pub struct BehaviorConfig {
+    pub tri_layer: Option<[u8; 3]>,
+    pub one_shot: OneShotConfig,
+}
+
+/// Config for one shot behavior
+pub struct OneShotConfig {
+    pub timeout: Duration,
+}
+
+impl Default for OneShotConfig {
+    fn default() -> Self {
+        Self {
+            timeout: Duration::from_secs(1),
         }
     }
 }